refactor(home): extract widgets and add docs
