+++ lib/presentation/home/screens/home_screen.dart
@@ -2,19 +2,20 @@
 
 import 'package:dear_flutter/core/di/injection.dart';
 import 'package:dear_flutter/domain/entities/audio_track.dart';
-import 'package:dear_flutter/domain/entities/motivational_quote.dart';
 import 'package:dear_flutter/presentation/home/cubit/latest_music_cubit.dart';
 import 'package:dear_flutter/presentation/home/cubit/latest_music_state.dart';
 import 'package:dear_flutter/presentation/home/cubit/latest_quote_cubit.dart';
 import 'package:dear_flutter/presentation/home/cubit/latest_quote_state.dart';
 import 'package:dear_flutter/services/audio_player_handler.dart';
 import 'package:dear_flutter/domain/repositories/song_history_repository.dart';
+import 'package:dear_flutter/presentation/home/widgets/widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'dart:async';
 import 'package:audio_service/audio_service.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:shimmer/shimmer.dart';
+
+/// Home page showing the latest quote and music recommendation.
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -131,9 +132,9 @@ class _HomeScreenState extends State<HomeScreen> {
                 child: ListView(
                   padding: const EdgeInsets.all(16),
                   children: [
-                    const _QuoteSection(),
+                    const QuoteSection(),
                     const SizedBox(height: 24),
-                    _MusicSection(
+                    MusicSection(
                       onPlay: _playTrack,
                       loading: _loading,
                     ),
@@ -142,7 +143,7 @@ class _HomeScreenState extends State<HomeScreen> {
               ),
             ),
             if (_currentTrack != null)
-              _PlayerBar(
+              PlayerBar(
                 track: _currentTrack!,
                 isPlaying: _isPlaying,
                 isLoading: _loading,
@@ -158,250 +159,3 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 }
 
-// Widget for displaying quotes with shimmer loading
-class _QuoteSection extends StatelessWidget {
-  const _QuoteSection();
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<LatestQuoteCubit, LatestQuoteState>(
-      builder: (context, state) {
-        if (state.status == LatestQuoteStatus.loading) {
-          return _ShimmerQuoteCard();
-        }
-        if (state.status == LatestQuoteStatus.failure) {
-          return Center(
-            child: Text(state.errorMessage ?? 'Terjadi kesalahan'),
-          );
-        }
-        final quote = state.quote;
-        if (quote == null) return const SizedBox.shrink();
-        return _QuoteCard(quote: quote);
-      },
-    );
-  }
-}
-
-// Shimmer effect for quote card while loading
-class _ShimmerQuoteCard extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Shimmer.fromColors(
-      baseColor: Colors.grey[300]!,
-      highlightColor: Colors.grey[100]!,
-      child: Card(
-        child: ListTile(
-          leading: Icon(Icons.format_quote, color: Colors.grey[400]),
-          title: Container(height: 20.0, color: Colors.grey[400]),
-          subtitle: Container(height: 15.0, color: Colors.grey[300]),
-        ),
-      ),
-    );
-  }
-}
-
-// Widget for displaying music cards
-class _MusicSection extends StatelessWidget {
-  const _MusicSection({required this.onPlay, required this.loading});
-
-  final Future<void> Function(AudioTrack) onPlay;
-  final bool loading;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<LatestMusicCubit, LatestMusicState>(
-      builder: (context, state) {
-        if (state.status == LatestMusicStatus.loading) {
-          return _ShimmerMusicCard();
-        }
-        if (state.status == LatestMusicStatus.failure) {
-          return Center(
-            child: Text(state.errorMessage ?? 'Terjadi kesalahan'),
-          );
-        }
-        final track = state.track;
-        if (track == null) {
-          return const SizedBox.shrink();
-        }
-
-        return Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              'Rekomendasi Musik',
-              style: Theme.of(context).textTheme.headlineSmall,
-            ),
-            const SizedBox(height: 16),
-            _MusicCard(
-              track: track,
-              onTap: () => onPlay(track),
-              loading: loading,
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-// Shimmer effect for music card while loading
-class _ShimmerMusicCard extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Shimmer.fromColors(
-      baseColor: Colors.grey[300]!,
-      highlightColor: Colors.grey[100]!,
-      child: Card(
-        child: ListTile(
-          leading: Icon(Icons.music_note, color: Colors.grey[400]),
-          title: Container(height: 20.0, color: Colors.grey[400]),
-          subtitle: Container(height: 15.0, color: Colors.grey[300]),
-        ),
-      ),
-    );
-  }
-}
-
-// Music card widget that shows song info
-class _MusicCard extends StatelessWidget {
-  const _MusicCard({
-    required this.track,
-    required this.onTap,
-    required this.loading,
-  });
-
-  final AudioTrack track;
-  final VoidCallback onTap;
-  final bool loading;
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        leading: const Icon(Icons.music_note),
-        title: Text(track.title),
-        subtitle: Text(track.artist ?? ''),
-        onTap: onTap,
-        trailing: loading
-            ? const SizedBox(
-                width: 24,
-                height: 24,
-                child: CircularProgressIndicator(strokeWidth: 2),
-              )
-            : null,
-      ),
-    );
-  }
-}
-
-// Quote card widget for displaying quotes
-class _QuoteCard extends StatelessWidget {
-  const _QuoteCard({required this.quote});
-
-  final MotivationalQuote quote;
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        leading: const Icon(Icons.format_quote),
-        title: Text('"${quote.text}"'),
-        subtitle: Text(quote.author),
-        onTap: () => context.go('/quote', extra: quote),
-      ),
-    );
-  }
-}
-
-// Player bar widget for controlling playback
-class _PlayerBar extends StatelessWidget {
-  const _PlayerBar({
-    required this.track,
-    required this.isPlaying,
-    required this.isLoading,
-    required this.position,
-    required this.duration,
-    required this.onToggle,
-    required this.onSeek,
-  });
-
-  final AudioTrack track;
-  final bool isPlaying;
-  final bool isLoading;
-  final Duration position;
-  final Duration duration;
-  final VoidCallback onToggle;
-  final ValueChanged<double> onSeek;
-
-  @override
-  Widget build(BuildContext context) {
-    final max = duration.inMilliseconds.toDouble();
-    final value = position.inMilliseconds.clamp(0, max).toDouble();
-    String fmt(Duration d) {
-      final minutes = d.inMinutes.remainder(60).toString().padLeft(2, '0');
-      final seconds = d.inSeconds.remainder(60).toString().padLeft(2, '0');
-      final hours = d.inHours;
-      return hours > 0 ? '$hours:$minutes:$seconds' : '$minutes:$seconds';
-    }
-
-    return Material(
-      color: Theme.of(context).colorScheme.surfaceContainerHighest,
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Slider(
-            min: 0,
-            max: max > 0 ? max : 1,
-            value: value,
-            onChanged: onSeek,
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-            child: Row(
-              children: [
-                if (track.coverUrl != null)
-                  ClipRRect(
-                    borderRadius: BorderRadius.circular(4),
-                    child: Image.network(
-                      track.coverUrl!,
-                      height: 40,
-                      width: 40,
-                      fit: BoxFit.cover,
-                    ),
-                  )
-                else
-                  Container(height: 40, width: 40, color: Colors.grey),
-                const SizedBox(width: 12),
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Text(
-                        track.title,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                      Text(
-                        '${fmt(position)} / ${fmt(duration)}',
-                        style: Theme.of(context).textTheme.bodySmall,
-                      ),
-                    ],
-                  ),
-                ),
-                isLoading
-                    ? const SizedBox(
-                        width: 24,
-                        height: 24,
-                        child: CircularProgressIndicator(strokeWidth: 2),
-                      )
-                    : IconButton(
-                        icon: Icon(isPlaying ? Icons.pause : Icons.play_arrow),
-                        onPressed: onToggle,
-                      ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
