+++ lib/presentation/home/widgets/widgets.dart
@@ -1,0 +1,3 @@
+export 'quote_section.dart';
+export 'music_section.dart';
+export 'player_bar.dart';
